Add timeout for udev rule
The version of udev in wheezy kills rules that take longer than 10s. The
long timeout of 120s is to be on the safe side.

Rule processing in udev is synchronous. So no other rules are processed
while this rule is executing. This is intentional, live-guest could have
some race conditions.